Moved extra config params init to __init__
